use uint32_t type for estimated delay variable

+++ PBKDF2-Wrapper/PBKDF2Configuration.h
@@ -34,7 +34,7 @@ NSString * NSStringFromPBKDF2PseudoRandomFunction(PBKDF2PseudoRandomFunction pse
 - (instancetype)initWithPasswordLength:(NSUInteger)passwordLength
                       derivedKeyLength:(NSUInteger)derivedKeyLength
                             saltLength:(NSUInteger)saltLength
-          estimatedDelayInMilliseconds:(NSUInteger)delayInMilliseconds;
+          estimatedDelayInMilliseconds:(uint32_t)delayInMilliseconds;
 
 - (instancetype)initWithDerivedKeyLength:(NSUInteger)derivedKeyLength
                               saltLength:(NSUInteger)saltLength
