+++ PBKDF2-Wrapper/PBKDF2Defaults.h
@@ -10,7 +10,7 @@
 #import "PBKDF2Configuration.h"
 
 extern NSUInteger PBKDF2DefaultDerivedKeyLength;
-extern NSUInteger PBKDF2DefaultDelayInMilliseconds;
+extern uint32_t   PBKDF2DefaultDelayInMilliseconds;
 extern NSUInteger PBKDF2DefaultSaltLength;
 extern NSUInteger PBKDF2DefaultRounds;
 extern PBKDF2PseudoRandomFunction PBKDF2DefaultPseudoRandomFunction;
